+++ apps/common/models_generator.py
@@ -27,6 +27,6 @@ class GeneratedApp(BaseModel):
 
     def __str__(self):
         if self.user:
-            return f"{ str(self.user.username) } - {self.task_result}: {str(self.generated_at)}, REPO: {self.gh_repo}" 
+            return f"{self.id} - {self.backend}, user={ str(self.user.username) } - {self.task_result}: {str(self.task_id)}, REPO: {self.gh_repo}" 
         else:
-            return f"{ str(self.user_ip) } - {self.task_result}: {str(self.generated_at)}, REPO: {self.gh_repo}" 
+            return f"{self.id} - {self.backend}, IP={ str(self.user_ip) } - {self.task_result}: {str(self.task_id)}, REPO: {self.gh_repo}"